refactoring 1.3

+++ source/animation/animation.cpp
@@ -48,7 +48,7 @@ namespace ginger {
 
 			_currentTick = 1;
 		} else {
-			_currentTick += round(time);
+			_currentTick += (int) round(time);
 		}
 	}
